+++ tests/core/test_metric_result_integration.py
@@ -132,12 +132,27 @@ def test_result_metric_integration():
 
         assert epoch_log == {"b": cumulative_sum, "a_epoch": cumulative_sum}
 
+    result.minimize = torch.tensor(1.0)
+    result.extra = {}
     assert str(result) == (
-        "ResultCollection(True, cpu, {"
+        "ResultCollection("
+        "minimize=1.0, "
+        "{"
         "'h.a': ResultMetric('a', value=DummyMetric()), "
         "'h.b': ResultMetric('b', value=DummyMetric()), "
         "'h.c': ResultMetric('c', value=DummyMetric())"
         "})"
+    )
+    assert repr(result) == (
+        "{"
+        "True, "
+        "device(type='cpu'), "
+        "minimize=tensor(1.), "
+        "{'h.a': ResultMetric('a', value=DummyMetric()), "
+        "'h.b': ResultMetric('b', value=DummyMetric()), "
+        "'h.c': ResultMetric('c', value=DummyMetric()), "
+        "'_extra': {}}"
+        "}"
     )
 
 
@@ -332,3 +347,9 @@ def test_lightning_module_logging_result_collection(tmpdir, device):
         gpus=1 if device == "cuda" else 0,
     )
     trainer.fit(model)
+
+
+def test_result_collection_extra_reference():
+    """Unit-test to check that the `extra` dict reference is properly set."""
+    rc = ResultCollection(True)
+    assert rc.extra is rc["_extra"]